requirements.txt
appdirs==1.4.3
Django==1.10.3
django-filter==1.0.2
djangorestframework==3.4.7
Markdown==2.6.8
packaging==16.8
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
